add storyshots
