+++ storybook-testing/storyshots.test.js
@@ -1,0 +1,12 @@
+/**
+ * @jest-environment jsdom
+ */
+import path from 'path';
+import initStoryshots, { multiSnapshotWithOptions } from '@storybook/addon-storyshots';
+
+// create jest snapshot tests from each story
+initStoryshots({
+    storyKindRegex: /^((?!.*?Component).)*$/,
+    integrityOptions: { cwd: path.join(__dirname, 'stories', 'stories') },
+    test: multiSnapshotWithOptions()
+});
